Omit minified and intellisense JavaScript files
This is mainly aimed at jQuery via NuGet. If we just use Cassette's defaults, we
end up loading it in three times, one of which doesn't work outside of Visual
Studio. However, these files may actually be useful, so we shouldn't delete them
outright. Just make sure that we remember to include the non-minified versions.